move transform to own file; update test; fix docs
