+++ spec/kiba/extend/transforms/explode/rows_from_multival_field_spec.rb
@@ -1,0 +1,32 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+RSpec.describe Kiba::Extend::Transforms::Explode::RowsFromMultivalField do
+  let(:accumulator){ [] }
+  let(:test_job){ Helpers::TestJob.new(input: input, accumulator: accumulator, transforms: transforms) }
+  let(:result){ test_job.accumulator }
+
+  let(:input) do
+    [
+      {r1: 'a;b', r2: 'foo;bar'}
+    ]
+  end
+  
+  let(:transforms) do
+    Kiba.job_segment do
+      transform Explode::RowsFromMultivalField, field: :r1, delim: ';'
+    end
+  end
+  
+  let(:expected) do
+    [
+      { r1: 'a', r2: 'foo;bar' },
+      { r1: 'b', r2: 'foo;bar' }
+    ]
+  end
+
+  it 'transforms as expected' do
+    expect(result).to eq(expected)
+  end
+end
